refactor(be): move methods that are resposible to publish judge message
